+++ packages/bruno-app/src/components/ResponsePane/QueryResult/QueryResultPreview/XmlPreview/index.js
@@ -1,44 +1,8 @@
-import React, { useEffect, useMemo, useRef, useState } from 'react';
+import React, { useMemo, useState } from 'react';
 import ErrorBanner from 'ui/ErrorBanner';
+import { useChunkedReveal } from 'hooks/useChunkedReveal';
 import { isHttpUrl } from 'utils/url';
 import StyledWrapper from './StyledWrapper';
-
-const CHUNK_SIZE = 200;
-
-/*
- * A node with a large number of children (e.g. a wide array) renders them in chunks
- * instead of all at once. A sentinel div sits after the last rendered child, and scrolling
- * it into view reveals the next chunk - the same pattern as an infinite-scroll list.
- */
-const useChunkedReveal = (total) => {
-  const [visibleCount, setVisibleCount] = useState(Math.min(total, CHUNK_SIZE));
-  const [prevTotal, setPrevTotal] = useState(total);
-  if (total !== prevTotal) {
-    setPrevTotal(total);
-    setVisibleCount(Math.min(total, CHUNK_SIZE));
-  }
-
-  const sentinelRef = useRef(null);
-
-  useEffect(() => {
-    if (visibleCount >= total) return undefined;
-    const sentinel = sentinelRef.current;
-    if (!sentinel) return undefined;
-
-    const observer = new IntersectionObserver(
-      (entries) => {
-        if (entries.some((entry) => entry.isIntersecting)) {
-          setVisibleCount((count) => Math.min(count + CHUNK_SIZE, total));
-        }
-      },
-      { rootMargin: '200px' }
-    );
-    observer.observe(sentinel);
-    return () => observer.disconnect();
-  }, [visibleCount, total]);
-
-  return [visibleCount, sentinelRef];
-};
 
 // The expected "data" prop must be an XML string.
 export default function XmlPreview({ data, defaultExpanded = true, onLinkClick }) {
